Move piv.SW to submodule piv.util
